backend/requirements.txt
fastapi
uvicorn
asyncpg
python-multipart
Pillow
passlib
pydantic
jose